Get current request path from context

Return the URL path of the current HTTP request, as captured in the freshwater request context. If called outside of an active context, an error is raised.

This is primarily intended for use inside templates where the request context has been established.

Context is available when freshwater context middleware is active (installed automatically by api_csrf(), api_error_pages(), or api_freshwater()).

Usage

current_path()

See also