Register context-safe async evaluator

Registers an asynchronous evaluator for routes, allowing freshwater context to be propagated to mirai::mirai workers. This means that contextful helpers such as current_method, current_path, and current_query (among others) will work in async routes.

Usage

register_async_evaluator(force = FALSE, set_default = TRUE)

Arguments

force

whether to register the evaluator regardless of if it has been registered already

set_default

whether to set the default async evaluator to "freshwater"

Details

Context is not inherently portable across asynchronous contexts, this function creates a portable snapshot of the current context that is passed to a mirai worker.

Hooks are not applied to the async route, but may be provided to any associated then handlers. If error pages are installed on the main process, errors from the worker will be appropriately converted into freshwater error pages.

Requires the promises::promises, mirai::mirai, and mori::mori packages.

See also

Examples

register_async_evaluator()
#* @async
#* @get /async
function() {
 current_path()
}
#> function () 
#> {
#>     current_path()
#> }
#> <environment: 0x558ab4aa2940>