Register HTML Serialiser

Registers an HTML serialiser for plumber2::plumber2 that renders shiny tags and taglists via htmltools::doRenderTags(). This is preferrable over htmltools::renderTags() as often we want to be able to emit head tags which htmltools::htmltools attempts to hide for shiny usecases.

Usage

register_html_serialiser(force = FALSE)

Arguments

force

bypass checks and re-register the freshwater serialiser

Details

The freshwater serialiser safely falls back to the default plumber2 implementation for other classes of inputs.