Create a reusable HTML template

Evaluates code under the htmltools::withTags() environment, creating a closure that outputs either the template or specified fragment.

The closure can also accept slots, specified as named or bare symbols prior to the template expression. This allows for passing in other templates or HTML tags after template creation.

template(..., .envir = parent.frame())

fragment(..., name = NULL)

Arguments

...

content

.envir

the environment in which to evaluate the template

name

the name of the fragment

Examples

# Example Fragment Usage
page_main <- template(
    {
        div(
            h1("Dashboard"),
            fragment(p("Welcome back"), name = "content"),
            small("2026")
        )
    }
)

page_main(fragment="content")
#> <p>Welcome back</p>

# Template slots

details <- template(name, age, {
    nm <- sprintf("Hello, my name is: %s", name)
    old <- sprintf("I am %s years old.", age)
    div(
        p(nm),
        p(old)
    )
})

details("Jim", 30)
#> <div>
#>   <p>Hello, my name is: Jim</p>
#>   <p>I am 30 years old.</p>
#> </div>

# Templates and fragments can also be combined

card <- template(
    ttl, foot = NULL, {
        div(
            h2(ttl),
            fragment(div("Card body"), name="body"),
            if (!is.null(foot)) {
                fragment(
                    div(foot),
                    name = "footer"
                )
            }
        )
    }
)
card("Card Title")
#> <div>
#>   <h2>Card Title</h2>
#>   <div>Card body</div>
#> </div>
card("Card Title", fragment="body")
#> <div>Card body</div>
card("Card Title", "Footer text", fragment = "footer")
#> <div>Footer text</div>